backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
supabase==2.28.0
python-dotenv==1.2.2
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
httpx==0.28.1
anthropic==0.49.0
pypdf==5.1.0
python-multipart==0.0.20
python-docx==1.1.2
reportlab==4.2.5
